+++ src/tcNetwork/Properties/AssemblyInfo.cs
@@ -3,7 +3,7 @@ Trade Control Network
 Date: 1.5.2020
 Author: IAM
 
-Documentation on github.com/tradecontrol/tcnodecore
+Documentation on github.com/tradecontrol/tc-network
 
 Trade Control by Trade Control Ltd is licensed under GNU General Public License v3.0. 
 
@@ -27,5 +27,5 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: Guid("0cc6cad7-45c6-47c1-bd81-5fe4309624b7")]
 
-[assembly: AssemblyVersion("1.1.0.0")]
-[assembly: AssemblyFileVersion("1.1.0.0")]
+[assembly: AssemblyVersion("1.2.0.0")]
+[assembly: AssemblyFileVersion("1.2.0.0")]